backend/requirements.txt
fastapi
uvicorn
solana
solders
openai
python-dotenv
